refactor(diff-schedule): split files and address review feedback
- Drop the redundant "diff" prefix from every file in the folder
  (diff.ts -> computeDiff.ts, diffResolvers.ts -> resolvers.ts,
  diffHelpers.ts -> helpers.ts, and matching test files).
- Extract the shared type definitions into types.ts.
- Replace handleRow with collectNewArtists/applyStageResolution helpers
  inlined directly in the computeDiff loop.
- Add resolvers.test.ts covering buildIndexes, resolveArtists,
  resolveStage, computeTimes and findMatchingSet.
- Move test data builders below the tests; drop the "computeDiff:"
  prefix from test titles; drop an unnecessary comment in resolvers.ts.
