+++ supabase/functions/diff-schedule/helpers.ts
